+++ src/components/links/LinkButton.jsx
@@ -60,6 +60,7 @@ import getButtonClassNames from '../../themes/components/button'
  * @param {number} [props.tabIndex] - Tab index.
  * @param {string} [props.title] - Accessible title (overrides i18n).
  * @param {string} [props.tooltip] - Tooltip text (overrides i18n).
+ * @param {import('../../themes/components/tooltip').TooltipAlignment} [props.tooltipAlign] - Tooltip alignment ('start' | 'center' | 'end').
  * @param {string} [props.tooltipClassName] - Tooltip class name.
  * @param {import('../../themes/components/tooltip').TooltipColorValue} [props.tooltipColor] - Tooltip color.
  * @param {import('../../themes/components/tooltip').TooltipPosition} [props.tooltipPosition] - Tooltip position.
@@ -89,6 +90,7 @@ const LinkButton =
     tabIndex ,
     title ,
     tooltip ,
+    tooltipAlign ,
     tooltipClassName ,
     tooltipColor ,
     tooltipPosition ,
@@ -107,6 +109,7 @@ const LinkButton =
 
     return (
         <Tooltip
+            align     = { tooltipAlign }
             className = { tooltipClassName }
             color     = { tooltipColor }
             position  = { tooltipPosition }